// src/app/landing/model/public-activity-type.enum.ts
export enum PublicActivityType {
  SIGNED_PETITION = "SIGNED_PETITION",
  JOINED_GROUP = "JOINED_GROUP",
  CALLED_MEETING = "CALLED_MEETING",
  CALLED_VOTE = "CALLED_VOTE",
  CREATED_TODO = "CREATED_TODO",
  CREATED_GROUP = "CREATED_GROUP",
  CREATED_ALERT = "CREATED_ALERT",
  SENT_BROADCAST = "SENT_BROADCAST",
  CREATED_CAMPAIGN = "CREATED_CAMPAIGN"
}